名前をArea,Pageで統一　Templete作成
