+++ app/components/Organisms/video-area.tsx
