+++ app/components/Templetes/closing-page.tsx
@@ -1,0 +1,13 @@
+import ClosingFirstView from "../Organisms/closing-first-view";
+import ClosingScondView from "../Organisms/closing-second-view";
+
+const ClosingPage = () => {
+  return (
+    <>
+      <ClosingFirstView />
+      <ClosingScondView />
+    </>
+  );
+};
+
+export default ClosingPage;
