+++ app/components/Templetes/home-page.tsx
@@ -1,0 +1,13 @@
+import ThumbnailCardArea from "../Organisms/thumbnail-card-area";
+import FirstViewArea from "../Organisms/first-view-area";
+
+const HomePage = () => {
+  return (
+    <>
+      <FirstViewArea />
+      <ThumbnailCardArea />
+    </>
+  );
+};
+
+export default HomePage;
