+++ app/section4/page.tsx
@@ -1,4 +1,4 @@
-import VideoPage from "../components/Organisms/video-page";
+import VideoPage from "../components/Templetes/video-page";
 import { step4VideoPageData } from "../const/video-page-data";
 
 const Page = () => {
